requirements.txt
aniso8601==9.0.1
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.13
urllib3==1.26.4
Werkzeug==1.0.1
